Updated and working prediction scripts for landsat and sentinel
